refactor(rpc): extract process_request to service/dispatch.rs (#148 part 1)
The transport-neutral request dispatch core (envelope verify → JWT/DPoP →
Casbin → handle_request → signed response) lived in transport/zmtp_quic.rs
but has no ZMQ/quinn deps. Move it verbatim to service/dispatch.rs so the
ZMQ-specific remainder of zmtp_quic.rs can be deleted in #138 without
touching the dispatch core, and so the #136 cutover references it from a
clean home.

- New service/dispatch.rs: process_request (verbatim) + the
  EnvelopeVerification re-export (the type itself stays in envelope.rs).
- service/mod.rs re-exports process_request.
- Repointed all callers off transport::zmtp_quic::process_request:
  RequestLoop (service/zmq.rs — also drops its last zmtp_quic dep, now
  uses crate::envelope::EnvelopeVerification), LocalServiceBridge
  (iroh_rpc.rs), and zmtp_quic's own QuicRep + WebTransportServer
  internal calls. Fixed two stale rustdoc links.

Pure move, no behavior change: 261 rpc lib tests green; workspace check clean.

Remaining #148 survivors: TLS helpers → transport/tls.rs (clean, follows);
WebTransportServer → transport/web_transport_server.rs is BLOCKED — its
handle_wt_stream still uses ZMTP multipart framing (handle_stream_*_zmtp),
which per the epic can't be de-ZMTP'd until FE-2 reframes the browser wire.
IrohRequestProcessor→RequestProcessor de-iroh (deferred from #166) also TBD.
